Serialize form refactor
Refactored the serializeform for each form
