+++ src/main/java/com/emma/app/action/AttendanceAction.java
@@ -11,16 +11,12 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
-@WebServlet("/add-attendance")
+@WebServlet("/attendance-sheet")
 public class AttendanceAction extends BaseAction {
 
     public void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        Attendance attendance = new Attendance();
+        Attendance attendance = serializeForm(Attendance.class, req.getParameterMap());
         AttendanceBeanI attendanceBean = new AttendanceBean();
-
-        HttpSession httpSession = req.getSession();
-
-        serializeForm(attendance, req.getParameterMap());
 
         try {
             attendanceBean.addorUpdateAttendance(attendance, req);
@@ -31,6 +27,12 @@ public class AttendanceAction extends BaseAction {
 
         resp.sendRedirect("./attendance-sheet");
 
+
+    }
+    public void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
+        AttendanceBeanI attendanceBean = new AttendanceBean();
+        renderAttendanceSheetPage(req, resp, 1);
 
     }
 }
